ml/requirements-prod.txt
tensorflow-cpu
opencv-python-headless
numpy
flask
flask-cors
gunicorn
